Added configuration check class
