+++ src/qilib/configuration_helper/adapters/__init__.py
@@ -1,4 +1,5 @@
 from qilib.configuration_helper.adapters.common_instrument_adapter import CommonInstrumentAdapter
+from qilib.configuration_helper.adapters.common_config_check_instrument_adapter import CommonConfigCheckInstrumentAdapter
 from qilib.configuration_helper.adapters.ami430_instrument_adapter import AMI430InstrumentAdapter
 from qilib.configuration_helper.adapters.keithley_dmm6500_adapter import Keithley6500InstrumentAdapter
 from qilib.configuration_helper.adapters.spi_rack_instrument_adapter import SpiRackInstrumentAdapter
