+++ src/qilib/configuration_helper/adapters/uhfli_instrument_adapter.py
@@ -17,7 +17,7 @@ WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEM
 COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
-from typing import Any, Optional
+from typing import Optional
 
 import numpy as np
 from qcodes.instrument_drivers.ZI.ZIUHFLI import ZIUHFLI
@@ -40,6 +40,3 @@ class ZIUHFLIInstrumentAdapter(CommonInstrumentAdapter):
             if 'val_mapping' in values:
                 values.pop('val_mapping')
         return parameters
-
-    def _compare_config_values(self, config_value: Any, device_value: Any, parameter: str) -> bool:
-        pass
